requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
kagglehub
jupyter
streamlit
pillow